Feature: Implemented basic lane driving capability based on lesson's walkthrough
Basically, this was done by finishing the Controller Object in the twist_controller. I fixed the waypoint_updater/waypoint_updater
Node (Lines 55 - 60) and increased the LOOKAHEAD_WPS back to 200 because i choose to run the simulator natively.
The method decelerate_waypoints was added to allow the car to stop at the traffic lights smoothly.
A further fix was added in line 147-148 as the callback function was lacking the instruction to write to controller.stopline_wp_idx